Initial commit of run validation scripts
This is an updated version of the equivalent scripts from the 2022 CAsT
task. The updates are mainly to add support for the PTKB provenances
used in the iKAT 2023 run format, but I've also begun adding some extra
checks for other parts of the run file.

The set of tests from the CAsT version of the scripts have still to be
updated, I'll add those in a later commit.