Make the old "strict" mode mandatory
The old --strict flag was used to abort validation if any passage IDs
failed to validate. This should actually always happen, so there's no
point in making it an option.

This commit removes it as an argparse and function parameter in main.py,
and updates the tests to match the new function signatures. A couple of
now-redundant tests are also removed.